sv-comp/fm-tools/lib-fm-tools/python/pyproject.toml
# This file is part of lib-fm-tools, a library for interacting with FM-Tools files:
# https://gitlab.com/sosy-lab/benchmarking/fm-tools
#
# SPDX-FileCopyrightText: 2024 Dirk Beyer <https://www.sosy-lab.org>
#
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fm-tools"
requires-python = ">= 3.10"
dynamic = ["version"]
authors = [{ name = "Henrik Wachowitz", email = "henrik.wachowitz@ifi.lmu.de" }]
maintainers = [
    { name = "Henrik Wachowitz", email = "henrik.wachowitz@ifi.lmu.de" },
    { name = "Dirk Beyer", email = "dirk.beyer@sosy-lab.org" },
]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "tomli >= 2.0; python_version <= '3.10'",
    "pyyaml >= 6.0",
    "werkzeug >= 3.0",
    "httpx[http2] >= 0.26",
    "tqdm >= 4.66",
    "benchexec>=3.27",
    "pyyaml",
]

[tool.hatch.version]
path = "src/fm_tools/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/fm_tools"]


[tool.hatch.envs.test]
dependencies = [
    "coverage[toml]",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pyfakefs",
    "requests",
    "requests-mock",
]
scripts.test = "pytest --cov=src/fm_tools --cov-report term --cov-report=xml tests/ {args}"

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.ruff]
line-length = 120

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
